TheGitaInaBox/requirements.txt
flask==3.0.3
langchain==0.1.12
langchain-core==0.1.50
langchain-groq==0.1.6
google-generativeai==0.7.1
PyPDF2==3.0.1